+++ index.js
@@ -1,5 +1,5 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import MainPage from "./src/MainPage";
+import MainPage from "./src/components/MainPage";
 
 ReactDOM.render(<MainPage />, document.getElementById("root"));
